requirements.txt
boto3
filechunkio